requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
numpy
matplotlib
pytesseract
python-multipart
python-bidi
arabic-reshaper